+++ src/directives/formly-form.js
@@ -118,12 +118,14 @@ function formlyForm(formlyUsability, formlyWarn, $parse, formlyConfig, $interpol
 
     function onModelOrFormStateChange() {
       angular.forEach($scope.fields, function runFieldExpressionProperties(field, index) {
-        /*jshint -W030 */
         const model = field.model || $scope.model;
         field.runExpressions && field.runExpressions();
         if (field.hideExpression) { // can't use hide with expressionProperties reliably
           const val = model[field.key];
           field.hide = evalCloseToFormlyExpression(field.hideExpression, val, field, index);
+        }
+        if (field.extras && field.extras.validateOnModelChange && field.formControl) {
+          field.formControl.$validate();
         }
       });
     }